Guarded normal pdf calculation
If any part of the multivariate normal pdf runs into issue, the pdf is output as 0. First condition number of covariance is checked, then exponent is checked to be negative, then determinant is checked to be non-negative, then square root is checked to be positive. Only if all these checks are satisfied, the pdf is calculated.
An extra check is imposed when computing conditional probability. If somehow the denominator is 0 (which is highly unlikely as there's regularization by signal covariance), the conditional probability is output as 1. Hence usual linear bayes estimator is computed.